Split network.go into multiple files
To follow what we do with containers and images.
